Added more functions
Added maxContrastIntensity which significantly improves the ability to spot the landmines by eye. Also added a viewPointClouds feature which allows for quick checking through the point clouds in a point cloud cell array. Unfortunately, it looks like the padding in the max contrast adds outliers, so that part does need to be fixed.